refactor commands
